Adjusted tests a bit

+++ tests/TestFiles/MultCase.hs
@@ -1,6 +1,6 @@
 module MultCase where
 
-f :: Double -> Bool
+f :: Int -> Bool
 f x = case x * 2 of
             4 -> True
             2 -> False
